1.8 Rewrite deploy, init and rsync.

+++ cmd/deploy.bash
@@ -2,42 +2,47 @@
 # Use of this source code is governed by a BSD-style
 # license that can be found in the LICENSE file.
 
+#
+# Usage:
+#     dev deploy <prj> vagrant 192.168.56.70
+#
 cmd_main() {
-    declare wkdir
+    if [[ $# -lt 3 ]]; then
+        dev_info "Usage: dev deploy <prj> <user> <host>" >&2
+        return 1;
+    fi
 
-    [[ -z "$glb_argv" || "$glb_argv" == "dev" ]] && wkdir="$glb_wkdir" || wkdir="${cfg_projects[$glb_argv]}"
+    declare prj="$1" wkdir user="$2" host="$3"
+
+    [[ "$1" == "dev" ]] && wkdir="$glb_wkdir" || wkdir="${cfg_projects[$1]}"
     [ -z "$wkdir" ] && echo "Config error in $glb_base/etc/dev.conf" >&2 && return 1
-    [ "${#glb_run_nodes[@]}" -le 0 ] && echo "Nodes is empty" >&2 && return 1
 
-    declare cmd user host sudoer init_tar=/tmp/devinit.tar
-
-    cmd="mkdir -m 700 -p \$HOME/.ssh &&"
-    cmd="$cmd cat > \$HOME/.ssh/authorized_keys &&"
-    cmd="$cmd chmod 600 \$HOME/.ssh/authorized_keys"
+    declare cmd init_tar=/tmp/devinit.tar
 
     [ -f $init_tar ] && rm -f $init_tar
 
     cd $wkdir
+
     tar --exclude='./devinit.tar' \
-        --exclude='./*.sublime-*' \
+        --exclude='./src' \
         --exclude='./tags' \
-        --exclude='./var/log' \
-        --exclude='./var/run' \
-        --exclude='./var/tmp' \
+        --exclude='./usr' \
+        --exclude='./var' \
         -cf $init_tar .
 
-    for host in "${!glb_run_nodes[@]}"; do
-        sudoer="${glb_run_nodes[$host]}"
-        [ -z "$glb_user" ] && user="$sudoer" || user="$glb_user"
+    cmd="mkdir -m 700 -p \$HOME/.ssh &&"
+    cmd="$cmd cat >> \$HOME/.ssh/authorized_keys &&"
+    cmd="$cmd chmod 600 \$HOME/.ssh/authorized_keys"
+    cat $glb_ssh_key.pub | dev_ssh $user@$host "$cmd"
 
-        # public key, remotely write to a file using SSH
-        # It's will overwrite authorized_keys
-        cat $glb_ssh_key.pub | dev_ssh $user@$host "$cmd"
-        [ "$user" != "$sudoer" ] && dev_scp $glb_ssh_key $glb_ssh_key.pub $user@$host:\$HOME/.ssh
-        dev_scp $init_tar $glb_base/bin/devinstaller $user@$host:/tmp
-        dev_ssh $user@$host "/tmp/devinstaller $wkdir $sudoer"
-    done
+    dev_scp $glb_ssh_key $glb_ssh_key.pub $user@$host:\$HOME/.ssh
+    dev_scp $init_tar $user@$host:/tmp
+
+    cmd="mkdir -p \$HOME/$prj/var/log &&"
+    cmd="$cmd tar -xf /tmp/devinit.tar -C \$HOME/$prj &&"
+    cmd="$cmd rm -f /tmp/devinit.tar &&"
+    cmd="$cmd \$HOME/$prj/bin/$prj.bash init dev"
+    dev_ssh $user@$host "$cmd"
 
     rm -f $init_tar
 }
-
